feature(torch forecasting): created TorchForecastingModel class and made rnn and tcn model to inherit from it
